third_party/WebKit/LayoutTests/http/tests/serviceworker/request-end-to-end.html

<!DOCTYPE html>
<title>Service Worker: Request end-to-end</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var t = async_test('Request: end-to-end');
t.step(function() {
    var url = 'resources/request-end-to-end-worker.js';
    var scope = 'resources/blank.html';

    service_worker_unregister_and_register(t, url, scope)
      .then(t.step_func(function(registration) {
          return wait_for_update(t, registration);
        }))
      .then(t.step_func(onUpdateFound))
      .catch(unreached_rejection(t));

    function sendMessagePort(worker) {
        var messageChannel = new MessageChannel();
        worker.postMessage({port:messageChannel.port2}, [messageChannel.port2]);
        return messageChannel.port1;
    }

    function onUpdateFound(sw) {
        var port = sendMessagePort(sw);
        port.addEventListener('message', t.step_func(function(event) {
            onMessage(event);
        }), false);
        port.start();
        sw.addEventListener('statechange', t.step_func(function(event) {
            if (event.target.state == 'activated')
                onActive();
        }));
    }

    function onActive() {
        with_iframe(scope, t.step_func(function(frame) {}));
    }

    function onMessage(event) {
        assert_equals(
            event.data.url,
            location.href.substring(0, location.href.lastIndexOf('/') + 1) + scope,
            'request.url should be passed to onfetch event.');
        assert_equals(event.data.method, 'GET',
                      'request.method should be passed to onfetch event.');
        assert_equals(event.data.headers['user-agent'], navigator.userAgent,
                      'User-Agent header should be passed to onfetch event.')
        service_worker_unregister_and_done(t, scope);
    }
});
</script>
